+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+- added support for `bestEffort` option of VRPC
+
 ## [3.1.0] - May 18 2022
 
 - added capability for setting an "identity" for the underlying VRPC client